third_party/WebKit/LayoutTests/fast/css/style-sharing-inline-stylesheet-expected.txt

Test that inline stylesheet does not confuse style sharing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.getElementsByClassName("testclass")[1]).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

first div
second div
